requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-mail==0.10.0
werkzeug==3.0.3
apscheduler==3.10.4
python-dotenv==1.0.1
gunicorn
mysql-connector-python==8.2.0
